chore: forcing rename of folder for casing issues
